Move tests to contextual folders
