+++ Middleware/src/test/java/ca/concordia/encs/citydata/producers/GeometryProducerTest.java
@@ -1,4 +1,4 @@
-package ca.concordia.encs.citydata;
+package ca.concordia.encs.citydata.producers;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -10,7 +10,6 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static org.hamcrest.Matchers.containsString;
 
-import java.util.ArrayList;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,14 +19,8 @@ import org.springframework.context.annotation.ComponentScan;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 
-import com.google.gson.JsonArray;
-import com.google.gson.JsonObject;
-
 import ca.concordia.encs.citydata.core.configs.AppConfig;
-import ca.concordia.encs.citydata.producers.EnergyConsumptionProducer;
-import ca.concordia.encs.citydata.producers.GeometryProducer;
 import ca.concordia.encs.citydata.operations.MergeOperation;
-import ca.concordia.encs.citydata.operations.StringFilterOperation;
 import ca.concordia.encs.citydata.PayloadFactory;
 
 /***
